refactor: make visualizer into full library
I had issues getting the visualizer to import properly into the web
framework, so I had to go back and make this a full library. I did the
following to make this happen:

- Rename _visualizer.rs_ to _lib.rs_
- Move the `#![feature(test)]` flag to the top of _lib.rs_
- Move _src/main.rs_ to _src/bin/vis_cli.rs
- Move everything from _src/cli.rs_ to _src/bin/vis_cli.rs_
- Delete _src/cli.rs_
- Fix imports & warnings
- Fix doctests now that they are actually being compiled (currently
        these are not running)

The CLI can still be run:
```
cargo run --bin vis_cli -- --help
```
