+++ holoviz/src/bin/vis_cli.rs
@@ -1,13 +1,33 @@
-#![feature(test)]
+// mod cli;
+// mod visualizer;
+// use clap::Parser;
 
-mod cli;
-mod visualizer;
+use holoviz::{Point, Visualizer};
+
 use clap::Parser;
+use std::path::PathBuf;
 
-use visualizer::{Point, Visualizer};
+#[derive(Parser)]
+#[command(author, version, about, long_about = None)]
+#[command(propagate_version = true)]
+pub struct Args {
+    /// Input file: expects a .svg with circles
+    pub input_svg: PathBuf,
+
+    /// Output file: expects .svg
+    pub output_svg: PathBuf,
+
+    /// Animation duration in seconds
+    #[arg(default_value_t = 2.)]
+    pub duration: f32,
+
+    /// Animate the output
+    #[arg(short, long)]
+    pub animate: bool,
+}
 
 fn main() {
-    let args = cli::Args::parse();
+    let args = Args::parse();
 
     let viz = Visualizer::from_file(args.input_svg).expect("test SVG file not found");
